[BOM-989] fix: 챌린지 비공개 답글 개수 버그 해결 (#686)
* fix: 댓글 목록 조회 시 조회자에 따른 비공개 답글 개수 오류 해결

* test: 댓글 목록 조회 시 조회자에 따른 비공개 답글 개수 테스트 작성

* test: 댓글 목록 조회 시 조회자에 따른 비공개 답글 개수 테스트 작성

* refactor: 컨트롤러에서 memberId로 받도록 수정

* refactor: 구독 유무 판단 시 EXISTS 사용하도록 수정

+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/challenge/dto/response/ChallengeCommentResponse.java
@@ -41,6 +41,6 @@ public record ChallengeCommentResponse(
         boolean isLiked,
 
         @Schema(requiredMode = RequiredMode.REQUIRED)
-        int replyCount
+        long replyCount
 ) {
 }
